Update to use track #40 for game over screen

+++ src/Randomizer.Tools/RomGenerator.cs
@@ -10,18 +10,21 @@ namespace Randomizer.Tools;
 
 public class RomGenerator
 {
-    private const string SmRomPath = @"E:\SMZ3\Super Metroid (Japan, USA) (En,Ja).sfc";
-    private const string Z3RomPath = @"E:\SMZ3\Zelda no Densetsu - Kamigami no Triforce (Japan).sfc";
+    private const string SmRomPath = @"D:\Games\SMZ3\Super Metroid (Japan, USA) (En,Ja).sfc";
+    private const string Z3RomPath = @"D:\Games\SMZ3\Zelda no Densetsu - Kamigami no Triforce (Japan).sfc";
     private const string IpsPath = @"D:\Source\SMZ3Randomizer\alttp_sm_combo_randomizer_rom\build\zsm.ips";
-    private const string OutputPath = @"E:\SMZ3\TestRom";
+    private const string OutputPath = @"D:\Games\SMZ3\TestRom";
     private const string PatchBatFile = @"D:\Source\SMZ3Randomizer\alttp_sm_combo_randomizer_rom\build.bat";
     private const string MsuPath = @"E:\SMZ3\MSUs\RetroPC\RPC-LTTP-MSU.msu";
     private const string RomName = "test-rom";
 
 
 
-    public static string GenerateRom(bool openRom)
+    public static string GenerateRom(string[] args)
     {
+        var openRom = args.Any(x => "openrom".Equals(x, StringComparison.OrdinalIgnoreCase));
+        var copyMsu = args.Any(x => "copymsu".Equals(x, StringComparison.OrdinalIgnoreCase));
+
         var process = Process.Start(new ProcessStartInfo()
         {
             FileName = PatchBatFile,
@@ -53,22 +56,25 @@ public class RomGenerator
         Rom.UpdateChecksum(rom);
         File.WriteAllBytes(romPath, rom);
 
-        foreach (var file in Directory.GetFiles(OutputPath))
+        if (copyMsu)
         {
-            if (file.EndsWith(".pcm") || file.EndsWith(".msu"))
+            foreach (var file in Directory.GetFiles(OutputPath))
             {
-                File.Delete(file);
+                if (file.EndsWith(".pcm") || file.EndsWith(".msu"))
+                {
+                    File.Delete(file);
+                }
             }
-        }
 
-        if (!string.IsNullOrEmpty(MsuPath))
-        {
-            File.Copy(MsuPath, Path.Combine(OutputPath, $"{RomName}.msu"));
-            for (var i = 1; i < 200; i++)
+            if (!string.IsNullOrEmpty(MsuPath))
             {
-                var pcmPath = MsuPath.Replace(".msu", $"-{i}.pcm");
-                if (File.Exists(pcmPath))
-                    NativeMethods.CreateHardLink(Path.Combine(OutputPath, $"{RomName}-{i}.pcm"), pcmPath, IntPtr.Zero);
+                File.Copy(MsuPath, Path.Combine(OutputPath, $"{RomName}.msu"));
+                for (var i = 1; i < 200; i++)
+                {
+                    var pcmPath = MsuPath.Replace(".msu", $"-{i}.pcm");
+                    if (File.Exists(pcmPath))
+                        NativeMethods.CreateHardLink(Path.Combine(OutputPath, $"{RomName}-{i}.pcm"), pcmPath, IntPtr.Zero);
+                }
             }
         }
